Add user reputation and views page and fix some other stuff on other pages

+++ StackOverflow_Statistics/Dtos/UsersCommentsCountDto.cs
@@ -2,7 +2,7 @@
 {
     public class UsersCommentsCountDto
     {
-        public int Position { get; set; }
+        public long Position { get; set; }
         public int? Id { get; set; }
         public string DisplayName { get; set; }
         public int? CommentCount { get; set; }
